Add the interface GameService
This interface contain all the request to deal with server.
Here we use `@GET` annotation for a GET request with the parameter `/scores`
According to the server this GET request return a list of score
 in JSON format.
Then after GET annotation we declare a function listScores without
params, because the request no need params.
The function listScores return an object Call of List<Score>,
Every function related to a request return a Call object, then the Call object take in <> what the server return for the related request.