Add The Score class
The Score class represent a Score object according to the server.
That means the name of constructor attribute of exactly same name of
attribute in JSON file returned by the server (pseudo and score).
It's possible to avoid some attribute form server like date in this way
we get only what we want.